+++ ElvUI_EltreumUI/Core/Options.lua
@@ -3725,6 +3725,7 @@ The Item Level shown on the Character Panel Skin uses code from Simple Item leve
 	ElvUI_EltreumUI.Options.args.changelog.args.added.inline = true
 	ElvUI_EltreumUI.Options.args.changelog.args.added.args.description = ACH:Description([[
 Added support for The War Within
+Added a menu in Eltruism > Installer to import and export Gradient settings
 Added several more statusbar textures from Blizzard's default textures
 Added the experimental Thin Mode to Eltruism > Installer > Layouts
 Added Quest and Delve Instances to the Instance Indicator
@@ -3751,6 +3752,7 @@ Fixed Pet Autocast error in Classic/Cataclysm due to 11.0 changes
 Fixed icon for professions in Classic and Cataclysm
 Fixed a possible error with the Quest Tracker skin
 Fixed a possible issue with the Quest Hide
+Fixed Hekili's skin hiding the Textures
 Fixed Event Toast lines
 Fixed Game Menu Button
 ]], 7, "small", nil, nil, nil, nil, "full")